+++ scenarios/standard/actions.py
@@ -3,7 +3,7 @@ import json
 import importlib
 
 from app import core
-from app.region_new import Region
+from app.region import Region
 from app.nationdata import Nation, NationTable
 from app.notifications import Notifications
 from app.war import Wars
